fix: handle AuthProviders components
