+++ apps/docs/content/_partials/providers.mdx
@@ -4,7 +4,7 @@ Supabase Auth works with many popular Auth methods, including Social and Phone A
 
 ### Social Auth
 
-<AuthProviders type="social" />
+<SocialAuthProviders />
 
 <Admonition type="tip">
 
@@ -14,4 +14,4 @@ You can also add any OAuth2 or OIDC-compatible identity provider using [Custom O
 
 ### Phone Auth
 
-<AuthProviders type="phone" />
+<PhoneAuthProviders />
